indices unsigned int

+++ MTV3D/VisComponent.h
@@ -78,7 +78,7 @@ private:
 	std::vector<Point> visPoints;
 	std::vector<std::vector<std::vector<VisComponent::Point>>> vis3DDataModel;
 	std::vector<Vertex> vertices;
-	std::vector<unsigned short> indices;
+	std::vector<unsigned> indices;
 
 	Microsoft::WRL::ComPtr<ID3D11Device> d3dDevice;
 	Microsoft::WRL::ComPtr<ID3D11DeviceContext> d3dDeviceContext;
